+++ src/js/Interactor/index.js
@@ -8,17 +8,7 @@ import UserStorage from '../UserStorage/index.js';
 
 import { checkKeyword } from '../Validator/index.js';
 
-export const formatDate = (publishTime) => {
-  const videoTime = new Date(publishTime.slice(0, 10));
-  return (
-    videoTime.getFullYear().toString() +
-    '년 ' +
-    (videoTime.getMonth() + 1).toString().padStart(2, '0') +
-    '월 ' +
-    videoTime.getDate().toString().padStart(2, '0') +
-    '일'
-  );
-};
+import { formatDate } from '../utils/index.js';
 
 const youtubeAPI = new YoutubeAPI();
