+++ src/js/utils/index.js
@@ -1,1 +1,13 @@
 export const isEmpty = (value) => value === '';
+
+export const formatDate = (publishTime) => {
+  const videoTime = new Date(publishTime.slice(0, 10));
+  return (
+    videoTime.getFullYear().toString() +
+    '년 ' +
+    (videoTime.getMonth() + 1).toString().padStart(2, '0') +
+    '월 ' +
+    videoTime.getDate().toString().padStart(2, '0') +
+    '일'
+  );
+};
